requirements.txt
# Context Engineer Core Dependencies
click>=8.1.0
jinja2>=3.1.2
pyyaml>=6.0.1
jsonschema>=4.0.0
numpy>=1.24.0  # For semantic similarity calculations
rich>=13.0.0  # For progress bars and beautiful terminal output

# Optional: For encrypted API key storage (pip install context-engineer[security])
# cryptography>=41.0.0

# Optional: For semantic embeddings (pip install context-engineer[ai])
# sentence-transformers>=2.2.2
# torch>=2.0.0

# Optional: For LLM provider SDKs (pip install context-engineer[providers])
# openai>=1.0.0
# anthropic>=0.3.0
# google-genai>=0.1.0
